# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminSummary {
  totalDevices: number;
  onlineDevices: number;
  pendingDevices: number;
  pausedDevices: number;
  commandsToday: number;
  /** @nullable */
  lastSyncAt: string | null;
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  DEVICE_ENROLLED: 'DEVICE_ENROLLED',
  DEVICE_APPROVED: 'DEVICE_APPROVED',
  DEVICE_DISABLED: 'DEVICE_DISABLED',
  COMMAND_SENT: 'COMMAND_SENT',
  DEVICE_CONNECTED: 'DEVICE_CONNECTED',
  DEVICE_DISCONNECTED: 'DEVICE_DISCONNECTED',
} as const;

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  message: string;
  /** @nullable */
  deviceId: string | null;
  /** @nullable */
  deviceName: string | null;
  createdAt: string;
}

export type DeviceStatus = typeof DeviceStatus[keyof typeof DeviceStatus];


export const DeviceStatus = {
  ONLINE: 'ONLINE',
  OFFLINE: 'OFFLINE',
  PAUSED: 'PAUSED',
  PENDING: 'PENDING',
} as const;

/**
 * @nullable
 */
export type DeviceActiveAction = typeof DeviceActiveAction[keyof typeof DeviceActiveAction] | null;


export const DeviceActiveAction = {
  NONE: 'NONE',
  BACK: 'BACK',
  HOME: 'HOME',
  NOTIFICATIONS: 'NOTIFICATIONS',
} as const;

export interface Device {
  id: string;
  displayName: string;
  model: string;
  androidVersion: string;
  appVersion: string;
  status: DeviceStatus;
  approved: boolean;
  /** @nullable */
  batteryLevel: number | null;
  /** @nullable */
  activeAction: DeviceActiveAction;
  /** @nullable */
  lastSeenAt: string | null;
  enrolledAt: string;
  updatedAt: string;
}

export interface DeviceEnrollmentInput {
  /** @minLength 1 */
  id: string;
  /** @minLength 1 */
  displayName: string;
  model: string;
  androidVersion: string;
  appVersion: string;
}

export interface DeviceEnrollmentResponse {
  device: Device;
  deviceToken: string;
}

export interface DeviceUpdate {
  /** @minLength 1 */
  displayName?: string;
  approved?: boolean;
  paused?: boolean;
}

export type DeviceCommandInputType = typeof DeviceCommandInputType[keyof typeof DeviceCommandInputType];


export const DeviceCommandInputType = {
  START: 'START',
  STOP: 'STOP',
  PAUSE: 'PAUSE',
  REFRESH_STATUS: 'REFRESH_STATUS',
} as const;

export type DeviceCommandInputAction = typeof DeviceCommandInputAction[keyof typeof DeviceCommandInputAction];


export const DeviceCommandInputAction = {
  NONE: 'NONE',
  BACK: 'BACK',
  HOME: 'HOME',
  NOTIFICATIONS: 'NOTIFICATIONS',
} as const;

export interface DeviceCommandInput {
  type: DeviceCommandInputType;
  action?: DeviceCommandInputAction;
  /** @nullable */
  templateName?: string | null;
}

export type DeviceCommandType = typeof DeviceCommandType[keyof typeof DeviceCommandType];


export const DeviceCommandType = {
  START: 'START',
  STOP: 'STOP',
  PAUSE: 'PAUSE',
  REFRESH_STATUS: 'REFRESH_STATUS',
} as const;

export type DeviceCommandAction = typeof DeviceCommandAction[keyof typeof DeviceCommandAction];


export const DeviceCommandAction = {
  NONE: 'NONE',
  BACK: 'BACK',
  HOME: 'HOME',
  NOTIFICATIONS: 'NOTIFICATIONS',
} as const;

export type DeviceCommandStatus = typeof DeviceCommandStatus[keyof typeof DeviceCommandStatus];


export const DeviceCommandStatus = {
  QUEUED: 'QUEUED',
  DELIVERED: 'DELIVERED',
  ACKNOWLEDGED: 'ACKNOWLEDGED',
  FAILED: 'FAILED',
  EXPIRED: 'EXPIRED',
} as const;

export interface DeviceCommand {
  id: string;
  deviceId: string;
  type: DeviceCommandType;
  action: DeviceCommandAction;
  /** @nullable */
  templateName: string | null;
  status: DeviceCommandStatus;
  createdAt: string;
  /** @nullable */
  acknowledgedAt: string | null;
}

/**
 * @nullable
 */
export type DeviceHeartbeatInputActiveAction = typeof DeviceHeartbeatInputActiveAction[keyof typeof DeviceHeartbeatInputActiveAction] | null;


export const DeviceHeartbeatInputActiveAction = {
  NONE: 'NONE',
  BACK: 'BACK',
  HOME: 'HOME',
  NOTIFICATIONS: 'NOTIFICATIONS',
} as const;

export interface DeviceHeartbeatInput {
  model: string;
  androidVersion: string;
  appVersion: string;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  batteryLevel: number | null;
  /** @nullable */
  activeAction: DeviceHeartbeatInputActiveAction;
  isRunning: boolean;
}

export interface HeartbeatResponse {
  device: Device;
  commands: DeviceCommand[];
}

export type DeviceTokenParameter = string;

export type GetActivityParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};

export type ListDevicesParams = {
status?: ListDevicesStatus;
search?: string;
};

export type ListDevicesStatus = typeof ListDevicesStatus[keyof typeof ListDevicesStatus];


export const ListDevicesStatus = {
  ALL: 'ALL',
  ONLINE: 'ONLINE',
  OFFLINE: 'OFFLINE',
  PAUSED: 'PAUSED',
  PENDING: 'PENDING',
} as const;