# lib/api-zod/src/generated/types/activityEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityEventType } from './activityEventType';

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  message: string;
  /** @nullable */
  deviceId: string | null;
  /** @nullable */
  deviceName: string | null;
  createdAt: Date;
}