# lib/api-zod/src/generated/types/deviceCommand.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceCommandAction } from './deviceCommandAction';
import type { DeviceCommandStatus } from './deviceCommandStatus';
import type { DeviceCommandType } from './deviceCommandType';

export interface DeviceCommand {
  id: string;
  deviceId: string;
  type: DeviceCommandType;
  action: DeviceCommandAction;
  /** @nullable */
  templateName: string | null;
  status: DeviceCommandStatus;
  createdAt: Date;
  /** @nullable */
  acknowledgedAt: Date | null;
}