# lib/api-zod/src/generated/types/deviceEnrollmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceEnrollmentInput {
  /** @minLength 1 */
  id: string;
  /** @minLength 1 */
  displayName: string;
  model: string;
  androidVersion: string;
  appVersion: string;
}