# lib/api-zod/src/generated/types/deviceHeartbeatInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceHeartbeatInputActiveAction } from './deviceHeartbeatInputActiveAction';

export interface DeviceHeartbeatInput {
  model: string;
  androidVersion: string;
  appVersion: string;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  batteryLevel: number | null;
  /** @nullable */
  activeAction: DeviceHeartbeatInputActiveAction;
  isRunning: boolean;
}