# lib/api-zod/src/generated/types/deviceUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceUpdate {
  /** @minLength 1 */
  displayName?: string;
  approved?: boolean;
  paused?: boolean;
}