# lib/api-zod/src/generated/types/heartbeatResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Device } from './device';
import type { DeviceCommand } from './deviceCommand';

export interface HeartbeatResponse {
  device: Device;
  commands: DeviceCommand[];
}